+++ packages/server/src/services/category/CategoryOrchestrationService.ts
@@ -3,30 +3,44 @@ import { ErrorCode, HttpCode, Utils } from '@tenpercent/shared';
 import { LoggerBase } from 'helper/logger/LoggerBase';
 import { IDatabaseConnection, IDBTransaction } from 'interfaces/IDatabaseConnection';
 import { ICategoryService } from 'services/category/CategoryService';
+import { GroupOrchestrationService } from 'services/groupOrchestrator/GroupOrchestrationService';
 import { ITransactionService } from 'services/transaction/TransactionService';
 import DatabaseConnectionBuilder from 'src/repositories/DatabaseConnectionBuilder';
 import { UnitOfWork } from 'src/repositories/UnitOfWork';
 import { CustomError } from 'src/utils/errors/CustomError';
+import { ValidationError } from 'src/utils/errors/ValidationError';
 import { AccountType } from 'types/AccountType';
 
 export class CategoryOrchestrationService extends LoggerBase {
     private readonly _categoryService: ICategoryService;
     private readonly _transactionService: ITransactionService;
+    private readonly _groupOrchestrationService: GroupOrchestrationService;
     constructor({
         categoryService,
         transactionService,
+        groupOrchestrationService,
     }: {
         categoryService: ICategoryService;
         transactionService: ITransactionService;
+        groupOrchestrationService: GroupOrchestrationService;
     }) {
         super();
         this._categoryService = categoryService;
         this._transactionService = transactionService;
+        this._groupOrchestrationService = groupOrchestrationService;
     }
 
     public async delete(userId: number, categoryId: number, keepData: boolean): Promise<boolean> {
         return this.withTransaction(async (trx: IDBTransaction) => {
             try {
+                const sharedItemId = await this._groupOrchestrationService.getSharedEntity(userId, 'categories', categoryId, trx);
+                if (Utils.isNotNull(sharedItemId)) {
+                    throw new ValidationError({
+                        message: `Delete category failed due reason: category ${categoryId} is shared with sharedItemId: ${sharedItemId}`,
+                        errorCode: ErrorCode.CATEGORY_DELETE_GROUP_ERROR,
+                        statusCode: HttpCode.BAD_REQUEST,
+                    });
+                }
                 if (!keepData) {
                     await this._transactionService.deleteTransactionsForEntity(
                         userId,
